--- a/parser/frontmatter.go
+++ b/parser/frontmatter.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/BurntSushi/toml"
 
-	yaml "gopkg.in/yaml.v2"
+	yaml "gopkg.in/yaml.v3"
 )
 
 const (