--- a/parser/metadecoders/decoder.go
+++ b/parser/metadecoders/decoder.go
@@ -28,7 +28,7 @@ import (
 	"github.com/spf13/afero"
 	"github.com/spf13/cast"
 	jww "github.com/spf13/jwalterweatherman"
-	yaml "gopkg.in/yaml.v2"
+	yaml "gopkg.in/yaml.v3"
 )
 
 // Decoder provides some configuration options for the decoders.