- [x] Update Hugo
- [ ] Update Cobra


The only failing (and now fixed tests) where some front matter with duplicate `weight` keys, which now is correctly reported as an error.

A random site benchmarks shows it to be a little bit heavier in the memor allocation department:

```yaml
name                                    old time/op    new time/op    delta
SiteNew/Regular_Tags_and_categories-16    43.2ms ± 4%    44.3ms ± 4%    ~     (p=0.200 n=4+4)

name                                    old alloc/op   new alloc/op   delta
SiteNew/Regular_Tags_and_categories-16    15.3MB ± 0%    15.5MB ± 0%  +1.46%  (p=0.029 n=4+4)

name                                    old allocs/op  new allocs/op  delta
SiteNew/Regular_Tags_and_categories-16      294k ± 0%      295k ± 0%  +0.24%  (p=0.029 n=4+4)
```

Fixes #7404